+++ pykube/config.py
@@ -76,6 +76,34 @@ class KubeConfig(object):
         self.filename = filename
         return self
 
+    @classmethod
+    def from_url(cls, url):
+        """
+        Creates an instance of the KubeConfig class from a single URL (useful
+        for interacting with kubectl proxy).
+        """
+        doc = {
+            "clusters": [
+                {
+                    "name": "self",
+                    "cluster": {
+                        "server": url,
+                    },
+                },
+            ],
+            "contexts": [
+                {
+                    "name": "self",
+                    "context": {
+                        "cluster": "self",
+                    },
+                }
+            ],
+            "current-context": "self",
+        }
+        self = cls(doc)
+        return self
+
     def __init__(self, doc):
         """
         Creates an instance of the KubeConfig class.
@@ -116,10 +144,11 @@ class KubeConfig(object):
         """
         if not hasattr(self, "_users"):
             us = {}
-            for ur in self.doc["users"]:
-                us[ur["name"]] = u = copy.deepcopy(ur["user"])
-                BytesOrFile.maybe_set(u, "client-certificate")
-                BytesOrFile.maybe_set(u, "client-key")
+            if "users" in self.doc:
+                for ur in self.doc["users"]:
+                    us[ur["name"]] = u = copy.deepcopy(ur["user"])
+                    BytesOrFile.maybe_set(u, "client-certificate")
+                    BytesOrFile.maybe_set(u, "client-key")
             self._users = us
         return self._users
 
@@ -152,7 +181,7 @@ class KubeConfig(object):
         """
         if self.current_context is None:
             raise exceptions.PyKubeError("current context not set; call set_current_context")
-        return self.users[self.contexts[self.current_context]["user"]]
+        return self.users.get(self.contexts[self.current_context].get("user", ""), {})
 
 
 class BytesOrFile(object):